refactor: add categories Разделил функции по категориям для документации
