+++ README.md
@@ -1,2 +1,9 @@
-# vk-js
-VK shared JS libs
+# @vkontakte/vkjs
+
+Набор общих функций для ВКонтакте
+
+```sh
+yarn add @vkontakte/vkjs
+```
+
+[Документация](https://vkcom.github.io/vkjs)
