+++ src/animation/index.ts
@@ -1,0 +1,9 @@
+export {
+  animate,
+  animationEvent,
+  transitionEvent,
+  waitAnimationEnd,
+  cancelWaitAnimationEnd,
+  waitTransitionEnd,
+  cancelWaitTransitionEnd,
+} from './animate';
