+++ src/array/index.ts
@@ -1,0 +1,10 @@
+export {
+  createArray,
+  sumArray,
+  averageArray,
+  uniqueArray,
+  shuffleArray,
+  chunkArray,
+  omitFromArray,
+  difference,
+} from './arrays';
