+++ src/other/clipboard.ts
