+++ src/other/detections.ts
