+++ src/other/getOffsetRect.ts
