+++ src/other/getPhotoSize.ts
