+++ src/other/random.ts
