+++ src/typecheck/index.ts
@@ -1,0 +1,11 @@
+export {
+  isObjectLike,
+  isArray,
+  isObject,
+  isUndefined,
+  isFunction,
+  isFormData,
+  isString,
+  isNumber,
+  isPromiseLike,
+} from './type_checkers';
